moved panda-related functions to panda_utils
